+++ share/model/dao/reference/MarginDetailSZ.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Float, Date, String
+from sqlalchemy import Column, Integer, Float, Date, String, BigInteger
 from share.model.dao import Base
 from share.util.numberUtil import toFloat, toInt
 from dateutil.parser import parse
@@ -40,11 +40,11 @@ class Model(Base):
     opDate = Column(Date, primary_key=True)  # 信用交易日期
     code = Column(String(32), primary_key=True)  # 标的证券代码
     name = Column(String(32))  # 标的证券简称
-    rzmre = Column(Float)  # 融资买入额(元)
-    rzye = Column(Float)  # 融资余额(元)
-    rqmcl = Column(Float)  # 融券卖出量
-    rqyl = Column(Float)  # 融券余量
-    rqye = Column(Float)  # 融券余量(元)
-    rzrqye = Column(Float)  # 融资融券余额(元)
+    rzmre = Column(BigInteger)  # 融资买入额(元)
+    rzye = Column(BigInteger)  # 融资余额(元)
+    rqmcl = Column(BigInteger)  # 融券卖出量
+    rqyl = Column(BigInteger)  # 融券余量
+    rqye = Column(BigInteger)  # 融券余量(元)
+    rzrqye = Column(BigInteger)  # 融资融券余额(元)
